+++ src/reducers/rampReducer.ts
